+++ lib/pie_chart_sz.dart
@@ -2,6 +2,7 @@ library pie_chart_sz;
 
 import 'package:flutter/material.dart';
 import 'package:pie_chart_sz/SpacedDonutChar.dart';
+import 'package:pie_chart_sz/ValueSettings.dart';
 
 class PieChartSz extends StatelessWidget {
   List<Color>? colors = [
@@ -20,14 +21,27 @@ class PieChartSz extends StatelessWidget {
     5,
   ];
   String? centerText;
-   PieChartSz({super.key,this.values,this.colors,this.centerText="center"});
+  Valuesettings? valueSettings;
+  double gapSize = 0.2;
+
+  PieChartSz(
+      {super.key,
+      this.values,
+      this.colors,
+      this.centerText = "Rezgar",
+      this.gapSize = 0.2,
+      this.valueSettings = const Valuesettings(showValues: false)});
 
   @override
   Widget build(BuildContext context) {
     return CustomPaint(
-    size: Size(300, 300),
-    painter: SpacedDonutChartPainter(colors:this.colors!,values: this.values!),
+      size: Size(300, 300),
+      painter: SpacedDonutChartPainter(
+          colors: this.colors!,
+          values: this.values!,
+          gapSize: gapSize,
+          valueSettings: valueSettings!,
+          centerText: centerText),
     );
   }
 }
-