+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 1.0.0
+
 * feat: added value display
+* feat:added value Settings
+* feat: added gap size
 * update README
+
 ## 0.0.2
+
 * feat: added center text
+
 ## 0.0.1 - Initial Release